# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Alive — Ghost Worker Verification API",
        "description": "AI proof-of-life verification system for Nigerian government payroll.",
        "contact": {
            "email": "api@sentinelbloom.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/agents": {
            "get": {
                "tags": [
                    "Field Agents"
                ],
                "summary": "List all field agents",
                "operationId": "agentIndex",
                "responses": {
                    "200": {
                        "description": "Paginated agents"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Field Agents"
                ],
                "summary": "Register a new field agent",
                "operationId": "agentStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "agent_code",
                                    "name",
                                    "phone",
                                    "state",
                                    "lga"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "agent_code": {
                                        "type": "string",
                                        "example": "AGT-001"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Emeka Obi"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "08090001234"
                                    },
                                    "state": {
                                        "type": "string",
                                        "example": "Anambra"
                                    },
                                    "lga": {
                                        "type": "string",
                                        "example": "Awka South"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Agent registered"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/agents/{id}": {
            "get": {
                "tags": [
                    "Field Agents"
                ],
                "summary": "Get a single field agent",
                "operationId": "agentShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Agent detail"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Field Agents"
                ],
                "summary": "Update a field agent",
                "operationId": "agentUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "state": {
                                        "type": "string"
                                    },
                                    "lga": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Agent updated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/agents/{id}/dispatches": {
            "get": {
                "tags": [
                    "Field Agents"
                ],
                "summary": "List dispatches for a field agent",
                "operationId": "agentDispatches",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated dispatches"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/agents/{id}/update-location": {
            "post": {
                "tags": [
                    "Field Agents"
                ],
                "summary": "Update GPS location of a field agent",
                "operationId": "agentUpdateLocation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "lat",
                                    "lng"
                                ],
                                "properties": {
                                    "lat": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 6.5244
                                    },
                                    "lng": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 3.3792
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Location updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}